[Feat] #10 - FireBase 주입 및 DataSource 연동 완료

+++ app/src/main/java/com/example/zupzup/domain/models/StoreHeaderInfoModel.kt
@@ -1,7 +1,7 @@
 package com.example.zupzup.domain.models
 
 data class StoreHeaderInfoModel(
-    val storeID : Int,
+    val storeID : Long,
     val name : String,
     val openTime : String,
     val maximumSales : Int
